Fixed Java Shit

+++ src/com/qmp/rest/services/CrudRestBase.java
@@ -225,7 +225,7 @@ public abstract class CrudRestBase extends RestBase {
 	@GET
 	@Path("/all/{timestamp}/{cd}")
 	@Produces(MediaType.APPLICATION_JSON)
-	public String getModified(@PathParam("timestamp") int timestamp, @PathParam("cd") int cd) {
+	public String getModified(@PathParam("timestamp") int timestamp, @PathParam("cd") Integer cd) {
 		if(timestamp == 0)
 			return getAll(cd);
 		
@@ -242,7 +242,7 @@ public abstract class CrudRestBase extends RestBase {
 	@Path("/all/{timestamp}")
 	@Produces(MediaType.APPLICATION_JSON)
 	public String getModified(@PathParam("timestamp") int timestamp) {
-		return getModified(timestamp, (Integer) null);
+		return getModified(timestamp, null);
 	}
